Add updateMed server action and initial fetch logic
